requirements/tests.txt
coverage==3.6
flake8==1.7
django-discover-runner==1.0
